[FIX] Indexing issue caused by r52407 unsupported type
git-svn-id: https://svn.code.sf.net/p/tikiwiki/code/trunk@52422 b456876b-0849-0410-b77d-98878d47e9d5

+++ lib/core/Search/Elastic/Index.php
@@ -69,7 +69,7 @@ class Search_Elastic_Index implements Search_Index_Interface, Search_Index_Query
 			function ($entry) {
 				if ($entry instanceof Search_Type_Numeric) {
 					return array(
-						"type" => "numeric",
+						"type" => "float",
 						"fields" => array(
 							"sort" => array(
 								"type" => "float",
